* USE: Enabled using ACESFilm on r2.

+++ game/gamedata/shaders/r2/aces.h
@@ -1,4 +1,4 @@
-float4 ACESFilm(float4 x)
+half4 ACESFilm(half4 x)
 {
     float a = 2.51f;
     float b = 0.03f;
